+++ src/v1/OptionTypes/numerical/NumericalConversionRules.ts
@@ -31,28 +31,19 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { implementsToPrimitive, implementsToString } from "../../../Protocols";
-import { NumericalRules } from "../NumericalRules";
+import { ComposableFunction } from "../../Archetypes";
+import { FunctionPointerTable } from "../../SupportingTypes";
 
 /**
- * `DEFAULT_NUMERICAL_RULES` are the default rules that
- * {@link resolveNumerical} uses to convert different types to numbers.
+ * `NumericalConversionRules` describes the table that tells
+ * {@link resolveNumerical} how to convert any given type to a `number`.
+ *
+ * Use {@link DEFAULT_NUMERICAL_RULES} if you ever need to pass the
+ * default ruleset into a function.
  *
  * @category OptionTypes
  */
-export const DEFAULT_NUMERICAL_RULES: NumericalRules = {
-    boolean: (x: boolean) => x ? 1 : 0,
-    number: (x: number) => x,
-    object: (x: object) => {
-        if (implementsToPrimitive(x)) {
-            return Number(+x);
-        }
-
-        if (implementsToString(x)) {
-            return Number(x.toString());
-        }
-
-        return NaN;
-    },
-    string: (x: string) => Number(x),
-}
+export type NumericalConversionRules = FunctionPointerTable<
+    any,
+    ComposableFunction<any, number>
+>;