+++ CHANGELOG.md
@@ -31,8 +31,8 @@ The following changes have been completed, and will be included in the next tagg
 * Option Types
   * numerical (new type!)
     - added `numerical` option type
-    - added `NumericalRules` function pointer table
-    - added `DEFAULT_NUMERICAL_RULES`
+    - added `NumericalConversionRules` function pointer table
+    - added `DEFAULT_NUMERICAL_CONVERSION_RULES`
     - added `resolveNumerical()`
     - added `validateNumericalData()`
 * Protocols